Q&A-Chatbot-with-Groq/requirements.txt
langchain
langchain-community
langchain-core
langchain_huggingface
langchain-chroma
beautifulsoup4
bs4 
langchain-groq
ollama
python-dotenv
streamlit